pyproject.toml
[project]
name = "polyglot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.0",
    "pregex>=2.3.3",
    "pydantic>=2.11.9",
    "pyparsing>=3.2.5",
    "pyyaml>=6.0",
    "rich>=14.2.0",
]
